Improve !ruoka to account for weekends
After this patch `!ruoka` will attempt to resolve problems
when faced with days that do not have a food defined in the database.
Now if a date is specified (or implied by no arguments) that is on a
weekend the bot will attempt to resolve it by querying for the food on
the following Monday. If that however does not produce a result, the bot
will give up.